+++ packages/suite/src/components/chat/components/ConnectWallet.tsx
@@ -23,14 +23,16 @@ export const ConnectWallet = () => {
   return (
     <div className="gas-h-full gas-flex gas-items-center gas-justify-center">
       <div className="gas-flex gas-flex-col gas-items-center gas-justify-center gas-gap-2">
-        <div className="gas-text-center" style={{ marginBottom: 10, color: theme.text.muted }}>
+        <div
+          className="gas-font-family gas-text-center"
+          style={{ marginBottom: 10, color: theme.text.muted }}>
           No wallet found. <br />
           Please connect your wallet to continue.
         </div>
         {connectButton ? (
           connectButton
         ) : (
-          <div className="gas-text-center" style={{ color: theme.text.muted }}>
+          <div className="gas-font-family gas-text-center" style={{ color: theme.text.muted }}>
             No wallet connect method found.
           </div>
         )}