+++ packages/suite/src/components/messages/system/error.tsx
@@ -8,9 +8,11 @@ export const SystemErrorMessage = ({
   content: SystemErrorMessageContent['content'];
 }) => {
   return (
-    <p className="gas-text-sm" style={{ wordWrap: 'break-word', overflowWrap: 'break-word' }}>
+    <p
+      className="gas-font-family gas-text-sm"
+      style={{ wordWrap: 'break-word', overflowWrap: 'break-word' }}>
       <span
-        className="gas-text-red-500 gas-flex gas-items-center gas-space-x-2"
+        className="gas-font-family gas-text-red-500 gas-flex gas-items-center gas-space-x-2"
         style={{ wordWrap: 'break-word', overflowWrap: 'break-word' }}>
         <AlertCircle className="gas-w-4 gas-h-4 gas-opacity-50 gas-mr-2" />
         {content}