+++ packages/suite/src/components/messages/system/text.tsx
@@ -3,10 +3,10 @@ import { TextMessageContent } from '@getgrowly/core';
 export const TextMessage = ({ content }: { content: TextMessageContent['content'] }) => {
   return (
     <p
-      className="gas-text-sm"
+      className="gas-font-family gas-text-sm"
       style={{ margin: 0, wordWrap: 'break-word', overflowWrap: 'break-word' }}>
       <span
-        className="gas-flex gas-items-center gas-space-x-2"
+        className="gas-font-family gas-flex gas-items-center gas-space-x-2"
         style={{ wordWrap: 'break-word', overflowWrap: 'break-word' }}>
         {content}
       </span>