+++ packages/suite/src/components/settings/components/SettingsPanel.tsx
@@ -5,12 +5,10 @@ import { useSuiteSession } from '@/hooks/use-session';
 import { cn } from '@/lib/utils';
 import { text } from '@/styles/theme';
 import { Screen } from '@/types/screen';
-import { Address, Avatar, Identity, Name } from '@coinbase/onchainkit/identity';
 import { ArrowLeft } from 'lucide-react';
 
 export const SettingsPanel = () => {
   const { setScreen, user } = useSuiteSession();
-  console.log('user', user);
   return (
     <PanelLayout>
       <div
@@ -18,15 +16,7 @@ export const SettingsPanel = () => {
           'gas-flex gas-flex-col gas-items-center gas-justify-center gas-h-full gas-p-[50px] gas-space-y-4',
           text.base
         )}>
-        {user?.entities.walletAddress ? (
-          <Identity address={user.entities.walletAddress} hasCopyAddressOnClick={true}>
-            <Avatar />
-            <Name />
-            <Address />
-          </Identity>
-        ) : (
-          <ConnectWallet />
-        )}
+        {!user?.entities.walletAddress && <ConnectWallet />}
         <Button
           onClick={() => {
             setScreen(Screen.Home);