+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @getgrowly/dashboard
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @getgrowly/suite@1.0.11
+
 ## 0.1.12
 
 ### Patch Changes